pyproject.toml
[project]
name = "chest_segment"
version = "0.1.0"
description = ""
authors = [
  {name = "agalaida", email = "alexgalayda@proton.me"}
]
maintainers = [
  {name = "agalaida", email = "alexgalayda@proton.me"}
]
requires-python = ">=3.12"

dependencies = [
    "numpy>=2.2.5",
    "torch>=2.6.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.1",
    "pydantic-settings>=2.9.1",
    "kagglehub>=0.3.12",
    "pillow>=11.2.1",
    "joblib>=1.5.1",
    "pandas>=2.3.0",
    "torchvision>=0.22.1",
    "scikit-image>=0.25.2",
    "tensorboard>=2.19.0",
    "hydra-core>=1.3.2",
    "segmentation-models-pytorch>=0.5.0",
    "torchmetrics>=1.7.3",
    "albumentations>=2.0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chest_segment"]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.3",
    "matplotlib>=3.10.3",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "ruff"
]

[tool.ruff]
line-length = 79
target-version = "py312"
show-fixes = true
src = ["src/chest_segment", "run.py"]

[tool.ruff.lint]
fixable = ["ALL", "I"]
ignore = [
  "D100",  # missing docstring in public module
  "D101", "D102", "D103", "D107",
  "ANN",        # missing type annotations
  "E501",       # line length
  "TRY003",     # long exception messages
  "EM101",      # string literal in exception
  "EM102",      # f-string in exception
]

select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.ruff.lint.isort]
known-first-party = ["chest_segment"] 